Condense business case sub-models for CEDAR Intake into one file
